Insert abstract class for common code of Software Module Details
