+++ hawkbit-ui/src/main/java/org/eclipse/hawkbit/ui/distributions/smtable/SwModuleDetails.java
@@ -8,120 +8,35 @@
  */
 package org.eclipse.hawkbit.ui.distributions.smtable;
 
-import org.apache.commons.lang3.StringUtils;
 import org.eclipse.hawkbit.repository.EntityFactory;
 import org.eclipse.hawkbit.repository.SoftwareManagement;
-import org.eclipse.hawkbit.repository.model.MetaData;
 import org.eclipse.hawkbit.repository.model.SoftwareModule;
 import org.eclipse.hawkbit.ui.SpPermissionChecker;
-import org.eclipse.hawkbit.ui.artifacts.event.SoftwareModuleEvent;
 import org.eclipse.hawkbit.ui.artifacts.smtable.SoftwareModuleAddUpdateWindow;
-import org.eclipse.hawkbit.ui.common.detailslayout.AbstractNamedVersionedEntityTableDetailsLayout;
-import org.eclipse.hawkbit.ui.common.detailslayout.SoftwareModuleMetadatadetailslayout;
-import org.eclipse.hawkbit.ui.components.SPUIComponentProvider;
-import org.eclipse.hawkbit.ui.distributions.event.MetadataEvent;
+import org.eclipse.hawkbit.ui.common.detailslayout.AbstractSoftwareModuleDetails;
 import org.eclipse.hawkbit.ui.distributions.state.ManageDistUIState;
-import org.eclipse.hawkbit.ui.utils.HawkbitCommonUtil;
 import org.eclipse.hawkbit.ui.utils.I18N;
 import org.eclipse.hawkbit.ui.utils.UIComponentIdProvider;
 import org.vaadin.spring.events.EventBus.UIEventBus;
-import org.vaadin.spring.events.EventScope;
-import org.vaadin.spring.events.annotation.EventBusListenerMethod;
-
-import com.vaadin.ui.Button.ClickEvent;
-import com.vaadin.ui.Label;
-import com.vaadin.ui.TabSheet;
-import com.vaadin.ui.UI;
-import com.vaadin.ui.VerticalLayout;
-import com.vaadin.ui.Window;
 
 /**
  * Implementation of software module details block using generic abstract
  * details style .
  */
-public class SwModuleDetails extends AbstractNamedVersionedEntityTableDetailsLayout<SoftwareModule> {
+public class SwModuleDetails extends AbstractSoftwareModuleDetails {
 
-    private static final long serialVersionUID = -1052279281066089812L;
-
-    private final SoftwareModuleAddUpdateWindow softwareModuleAddUpdateWindow;
+    private static final long serialVersionUID = 1L;
 
     private final ManageDistUIState manageDistUIState;
-
-    private final transient SoftwareManagement softwareManagement;
-
-    private final SwMetadataPopupLayout swMetadataPopupLayout;
-
-    private final SoftwareModuleMetadatadetailslayout swmMetadataTable;
 
     SwModuleDetails(final I18N i18n, final UIEventBus eventBus, final SpPermissionChecker permissionChecker,
             final SoftwareModuleAddUpdateWindow softwareModuleAddUpdateWindow,
             final ManageDistUIState manageDistUIState, final SoftwareManagement softwareManagement,
             final SwMetadataPopupLayout swMetadataPopupLayout, final EntityFactory entityFactory) {
-        super(i18n, eventBus, permissionChecker, null);
-        this.softwareModuleAddUpdateWindow = softwareModuleAddUpdateWindow;
+        super(i18n, eventBus, permissionChecker, null, softwareManagement, swMetadataPopupLayout, entityFactory,
+                softwareModuleAddUpdateWindow);
         this.manageDistUIState = manageDistUIState;
-        this.softwareManagement = softwareManagement;
-        this.swMetadataPopupLayout = swMetadataPopupLayout;
-
-        swmMetadataTable = new SoftwareModuleMetadatadetailslayout();
-        swmMetadataTable.init(getI18n(), getPermissionChecker(), softwareManagement, swMetadataPopupLayout,
-                entityFactory);
-        addTabs(detailsTab);
         restoreState();
-    }
-
-    /**
-     * MetadataEvent.
-     *
-     * @param event
-     *            as instance of {@link MetadataEvent}
-     */
-
-    @EventBusListenerMethod(scope = EventScope.UI)
-    void onEvent(final MetadataEvent event) {
-        UI.getCurrent().access(() -> {
-            final MetaData softwareModuleMetadata = event.getMetaData();
-            if (softwareModuleMetadata != null && isSoftwareModuleSelected(event.getModule())) {
-                if (event.getMetadataUIEvent() == MetadataEvent.MetadataUIEvent.CREATE_SOFTWARE_MODULE_METADATA) {
-                    swmMetadataTable.createMetadata(event.getMetaData().getKey());
-                } else if (event
-                        .getMetadataUIEvent() == MetadataEvent.MetadataUIEvent.DELETE_SOFTWARE_MODULE_METADATA) {
-                    swmMetadataTable.deleteMetadata(event.getMetaData().getKey());
-                }
-            }
-        });
-    }
-
-    @EventBusListenerMethod(scope = EventScope.UI)
-    void onEvent(final SoftwareModuleEvent softwareModuleEvent) {
-        onBaseEntityEvent(softwareModuleEvent);
-    }
-
-    @Override
-    protected void onEdit(final ClickEvent event) {
-        final Window addSoftwareModule = softwareModuleAddUpdateWindow
-                .createUpdateSoftwareModuleWindow(getSelectedBaseEntityId());
-        addSoftwareModule.setCaption(getI18n().get("upload.caption.update.swmodule"));
-        UI.getCurrent().addWindow(addSoftwareModule);
-        addSoftwareModule.setVisible(Boolean.TRUE);
-    }
-
-    @Override
-    protected String getEditButtonId() {
-        return UIComponentIdProvider.UPLOAD_SW_MODULE_EDIT_BUTTON;
-    }
-
-    @Override
-    protected void addTabs(final TabSheet detailsTab) {
-        detailsTab.addTab(createDetailsLayout(), getI18n().get("caption.tab.details"), null);
-        detailsTab.addTab(createDescriptionLayout(), getI18n().get("caption.tab.description"), null);
-        detailsTab.addTab(createLogLayout(), getI18n().get("caption.logs.tab"), null);
-        detailsTab.addTab(swmMetadataTable, getI18n().get("caption.metadata"), null);
-    }
-
-    @Override
-    protected String getDefaultCaption() {
-        return getI18n().get("upload.swModuleTable.header");
     }
 
     @Override
@@ -135,87 +50,13 @@ public class SwModuleDetails extends AbstractNamedVersionedEntityTableDetailsLay
     }
 
     @Override
-    protected Boolean hasEditPermission() {
-        return getPermissionChecker().hasUpdateDistributionPermission();
-    }
-
-    @Override
     protected String getTabSheetId() {
         return UIComponentIdProvider.DIST_SW_MODULE_DETAILS_TABSHEET_ID;
     }
 
-    private void populateDetails() {
-        if (getSelectedBaseEntity() != null) {
-            String maxAssign;
-            if (getSelectedBaseEntity().getType().getMaxAssignments() == 1) {
-                maxAssign = getI18n().get("label.singleAssign.type");
-            } else {
-                maxAssign = getI18n().get("label.multiAssign.type");
-            }
-            updateSwModuleDetailsLayout(getSelectedBaseEntity().getType().getName(),
-                    getSelectedBaseEntity().getVendor(), maxAssign);
-        } else {
-            updateSwModuleDetailsLayout(StringUtils.EMPTY, StringUtils.EMPTY, StringUtils.EMPTY);
-        }
-    }
-
-    private void updateSwModuleDetailsLayout(final String type, final String vendor, final String maxAssign) {
-
-        final VerticalLayout detailsTabLayout = getDetailsLayout();
-        detailsTabLayout.removeAllComponents();
-
-        final Label vendorLabel = SPUIComponentProvider.createNameValueLabel(getI18n().get("label.dist.details.vendor"),
-                HawkbitCommonUtil.trimAndNullIfEmpty(vendor) == null ? "" : vendor);
-        vendorLabel.setId(UIComponentIdProvider.DETAILS_VENDOR_LABEL_ID);
-        detailsTabLayout.addComponent(vendorLabel);
-
-        if (type != null) {
-            final Label typeLabel = SPUIComponentProvider.createNameValueLabel(getI18n().get("label.dist.details.type"),
-                    type);
-            typeLabel.setId(UIComponentIdProvider.DETAILS_TYPE_LABEL_ID);
-            detailsTabLayout.addComponent(typeLabel);
-        }
-
-        final Label assignLabel = SPUIComponentProvider.createNameValueLabel(getI18n().get("label.assigned.type"),
-                HawkbitCommonUtil.trimAndNullIfEmpty(maxAssign) == null ? "" : maxAssign);
-        assignLabel.setId(UIComponentIdProvider.SWM_DTLS_MAX_ASSIGN);
-        detailsTabLayout.addComponent(assignLabel);
-
-    }
-
     @Override
-    protected void populateDetailsWidget() {
-        populateDetails();
-        populateMetadataDetails();
+    protected boolean isSoftwareModuleSelected(final SoftwareModule softwareModule) {
+        return compareSoftwareModulesById(softwareModule, manageDistUIState.getSelectedBaseSwModuleId());
     }
 
-    @Override
-    protected String getDetailsHeaderCaptionId() {
-        return UIComponentIdProvider.TARGET_DETAILS_HEADER_LABEL_ID;
-    }
-
-    @Override
-    protected void populateMetadataDetails() {
-        swmMetadataTable.populateSMMetadata(getSelectedBaseEntity());
-    }
-
-    private boolean isSoftwareModuleSelected(final SoftwareModule softwareModule) {
-        if (softwareModule == null) {
-            return false;
-        }
-
-        return manageDistUIState.getSelectedBaseSwModuleId().map(module -> module.equals(softwareModule.getId()))
-                .orElse(false);
-    }
-
-    @Override
-    protected Boolean isMetadataIconToBeDisplayed() {
-        return true;
-    }
-
-    @Override
-    protected void showMetadata(final ClickEvent event) {
-        softwareManagement.findSoftwareModuleById(getSelectedBaseEntityId())
-                .ifPresent(swmodule -> UI.getCurrent().addWindow(swMetadataPopupLayout.getWindow(swmodule, null)));
-    }
 }
